start_bundle_coupled

Start coupled REMIND and MAGPIE runs

Usage

start_bundle_coupled(
  path_remind,
  path_magpie,
  path_settings_coupled,
  path_settings_remind,
  prefix_runname = "C_",
  path_remind_oldruns = paste0(path_remind, "output"),
  path_magpie_oldruns = paste0(path_magpie, "output"),
  prefix_oldruns = "C_",
  max_iterations = 5,
  n600_iterations = 0,
  test = FALSE
)

Arguments

path_remind

Path to the remind model

path_magpie

Path to the mapie model

path_settings_coupled

path_settings_coupled defines which runs will be started, coupling infos, and optimal gdx and report information that overrides path_settings_remind

path_settings_remind

path_settings_remind contains the detailed configuration of the REMIND scenarios

prefix_runname

"C_"

path_remind_oldruns

If there are existing runs you would like to take the gdxes (REMIND) or reportings (REMIND or MAgPIE) from, provide the path here and the name prefix below. Note: the scenario names of the old runs have to be identical to the runs that are to be started. If they differ please provide the names of the old scenarios in the file that you specified on path_settings_coupled (scenario_config_coupled_xxx.csv).

path_magpie_oldruns

If there are existing runs you would like to take the gdxes (REMIND) or reportings (REMIND or MAgPIE) from, provide the path here and the name prefix below. Note: the scenario names of the old runs have to be identical to the runs that are to be started. If they differ please provide the names of the old scenarios in the file that you specified on path_settings_coupled (scenario_config_coupled_xxx.csv).

prefix_oldruns

If you want the script to find gdxs or reports of older runs as starting point for new runs please provide the prefix of the old run names so the script can find them.

max_iterations

Integer

n600_iterations

Number of coupling iterations (before final iteration) in which MAgPIE uses higher n600 resolution. Until "max_iteration - n600_iterations" iteration MAgPIE runs with n200 resolution. Afterwards REMIND runs for "n600_iterations" iterations with results from higher resolution.

test

Logical. Set to TRUE to test function.

Examples

if (FALSE) {
# Start start_bundle_coupled runs
start_bundle_coupled("remind/", "magpie/", "remind/config/file1.csv", "remind/config/file2.csv")
}